requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# ── Database ─────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0,<3.0.0

# ── Speech-to-Text ──────────────────────────────────────────────────────
openai-whisper>=20231117

# ── Text-to-Speech ──────────────────────────────────────────────────────
gTTS>=2.4.0,<3.0.0

# ── Sentiment Analysis ──────────────────────────────────────────────────
transformers>=4.35.0,<5.0.0
torch>=2.1.0

# ── Audio Processing ────────────────────────────────────────────────────
ffmpeg-python>=0.2.0

# ── Utilities ────────────────────────────────────────────────────────────
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ── Testing ──────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# ── Optional: PostgreSQL driver (uncomment for production) ──────────────
# psycopg2-binary>=2.9.0
